requirements.txt
flask
python-escpos
pyusb
pillow
flask-cors
paho-mqtt
numpy
